Changing files to connect to our DB
I set up the DB and was able to connect to it via the login page. We now
have a fully working log in page!
I also created an update password page just to test queries. I was
successful in getting an update query to work and change the password.
This is good news, I now have a better idea of how PHP handles queries.