Assign and Link
I got the assigning portion working, as well as the generation of the
one-time-link. I've tested the link and it pulls the information we need
to display the specific rubric for that teacher that will be editing it.

All we need now is the display of the actual rubric, and the email
generation which I think Neil has worked on. We need to find out how to
use the "@selu.edu" type account.